add change

+++ response.go
@@ -2,7 +2,7 @@ package main
 
 type Response struct {
 	Success bool
-	Result  float32
+	Result  float64
 	Query   struct {
 		From   string
 		To     string
